move the functions from settings.py into helpers.py
